Databasemcp/requirements.txt
psycopg2-binary>=2.9.0
pyodbc>=4.0.30
